LocalStoryMap/pyproject.toml
[tool.poetry]
name = "localstorymap"
version = "0.1.0"
description = ""
authors = ["Alex424525 <sht914@gmail.com>"]

[[tool.poetry.packages]]
include = "apps"

[[tool.poetry.packages]]
include = "config"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = "^5.2.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
djangorestframework = "^3.16.0"
django-storages = "^1.14.6"
requests = "^2.32.3"
django-cors-headers = "^4.7.0"
redis = "^6.1.0"
django-redis = "^5.4.0"
gunicorn = "^23.0.0"
boto3 = "1.35.99"
sentry-sdk = "^2.29.1"
django-debug-toolbar = "^5.2.0"
black = "^23.0"
isort = "^5.0"
mypy = "^1.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.2.0"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.16.0"}
drf-yasg = "^1.21.10"
djangorestframework-simplejwt = "^5.5.0"
humps = "^0.2.2"
pyhumps = "^3.8.0"
drf-spectacular = "^0.28.0"
channels = "^4.2.2"
channels-redis = "^4.2.1"
pillow = "^11.2.1"
haversine = "^2.9.0"

[tool.poetry.group.dev.dependencies]
django-extensions = "^4.1"
drf-yasg = "^1.21.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
# Django·DRF 외에 타입 정보가 없는 서드파티 모듈은 무시
ignore_missing_imports = true

# django-stubs와 djangorestframework-stubs용 플러그인 로드
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[tool.django-stubs]
# Django 설정 파일 경로 지정 (개발용)
django_settings_module = "config.settings.dev"
strict_settings = true

[tool.black]
line-length = 88      # Black 기본값. 바꾼 게 있으면 여기에 맞춤

[tool.isort]
profile = "black"     # Black 스타일로 wrapping/줄바꿈 맞추기
line_length = 88      # Black과 동일하게